Digit Drawing + Tetris Finishing

+++ src/sk/estesadohodneme/sturik/game/Game.java
@@ -8,12 +8,12 @@ public abstract class Game {
 	/**
 	 * Colors for description of game board.
 	 */
-	public static final int COLOR_BLACK = 0;
-	public static final int COLOR_GREEN = 1;
-	public static final int COLOR_YELLOW= 2;
-	public static final int COLOR_RED   = 3;
-	public static final int COLOR_BLUE  = 4;
-	public static final int COLOR_WHITE = 5;
+	public static final short COLOR_BLACK = 0;
+	public static final short COLOR_GREEN = 1;
+	public static final short COLOR_YELLOW= 2;
+	public static final short COLOR_RED   = 3;
+	public static final short COLOR_BLUE  = 4;
+	public static final short COLOR_WHITE = 5;
 	
 	/**
 	 *
